+++ src/starlight_codec/__init__.py
@@ -3,19 +3,25 @@
 from .codec import (
     DecodeResult,
     EncodeResult,
+    CapsuleResult,
+    create_capsule_file,
     decode_file,
     decode_slb1,
     encode_file,
     encode_slb1,
+    hydrate_file,
     inspect_slb1,
 )
 
 __all__ = [
     "DecodeResult",
     "EncodeResult",
+    "CapsuleResult",
+    "create_capsule_file",
     "decode_file",
     "decode_slb1",
     "encode_file",
     "encode_slb1",
+    "hydrate_file",
     "inspect_slb1",
 ]